[dist] align to mocha framework
