--- packages/gasket-intl/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  extends: ['godaddy-react', 'plugin:jest/recommended'],
-  parser: "babel-eslint",
-  rules: {
-    "jest/valid-expect": "off"
-  }
-};
