--- packages/gasket-intl/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  setupFiles: ['./test/setup.js'],
-  collectCoverageFrom: ['src/**/*.js']
-};
