+++ packages/gasket-intl/package.json
@@ -6,13 +6,14 @@
   "browser": "./lib",
   "module": "./src",
   "scripts": {
-    "lint": "eslint src",
+    "lint": "eslint .",
     "lint:fix": "npm run lint -- --fix",
-    "test": "jest",
-    "test:watch": "jest --watchAll",
-    "test:coverage": "jest --coverage",
+    "test": "npm run test:runner",
+    "test:runner": "mocha --require setup-env --require test/setup.js \"test/*.test.js\"",
+    "test:watch": "npm run test:runner -- --watch",
+    "test:coverage": "nyc --reporter=text --reporter=json-summary npm run test:runner",
     "posttest": "npm run lint",
-    "build": "babel src -d lib --ignore \"**/*.spec.js\"",
+    "build": "babel src -d lib",
     "prepublishOnly": "npm run build"
   },
   "repository": {
@@ -49,7 +50,9 @@
     "@babel/plugin-transform-runtime": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
-    "acorn": "^6.1.1",
+    "@babel/register": "^7.5.5",
+    "assume": "^2.2.0",
+    "assume-sinon": "^1.0.1",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.2",
     "babel-plugin-root-import": "^5.1.0",
@@ -57,22 +60,35 @@
     "enzyme-adapter-react-16": "^1.14.0",
     "eslint": "^6.1.0",
     "eslint-config-godaddy-react": "^6.0.0",
-    "eslint-plugin-jest": "^22.15.1",
     "eslint-plugin-json": "^1.4.0",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-mocha": "^6.0.0",
     "eslint-plugin-react": "^7.14.0",
-    "jest": "^24.8.0",
+    "jsdom": "^16.4.0",
+    "mocha": "^6.2.0",
+    "nyc": "^14.1.1",
+    "proxyquire": "^2.1.3",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-intl": "^3.9.1",
     "react-redux": "^7.1.0",
     "redux": "^4.0.4",
-    "redux-mock-store": "^1.5.1"
+    "redux-mock-store": "^1.5.1",
+    "setup-env": "^1.2.2",
+    "sinon": "^7.4.1"
   },
   "peerDependencies": {
     "@gasket/plugin-intl": "^5.5.0",
     "react-intl": ">=2.0.0 <4.0.0",
     "react-redux": ">=5.0.0 <8.0.0"
-  }
+  },
+  "eslintConfig": {
+    "extends": [
+      "godaddy-react"
+    ]
+  },
+  "eslintIgnore": [
+    "lib",
+    "coverage"
+  ]
 }
