+++ packages/gasket-intl/test/fixtures/locales/en-US.json
@@ -1,0 +1,4 @@
+{
+  "gasket_welcome": "Hello!",
+  "gasket_learn": "Learn Gasket"
+}
