+++ packages/gasket-intl/test/fixtures/locales/fr-FR.json
@@ -1,0 +1,4 @@
+{
+  "gasket_welcome": "Bonjour!",
+  "gasket_learn": "Apprendre Gasket"
+}
