+++ packages/gasket-intl/test/fixtures/mock-manifest.json
@@ -1,0 +1,9 @@
+{
+  "basePath": "",
+  "defaultPath": "/locales",
+  "defaultLocale": "en-US",
+  "paths": {
+    "en-US.json": "10decbe",
+    "fr-FR.json": "21047f1"
+  }
+}
