+++ packages/gasket-intl/test/utils.test.js
@@ -1,23 +1,24 @@
+import assume from 'assume';
 import { formatLocalePath } from '../src/utils';
 
 describe('utils', function () {
 
-  describe('getLocaleFile', function () {
+  describe('formatLocalePath', function () {
     it('adds locale json to root path', function () {
       const results = formatLocalePath('/locales', 'en-US');
-      expect(results).toEqual('/locales/en-US.json');
+      assume(results).equals('/locales/en-US.json');
     });
     it('substitutes $locale in path template', function () {
       const results = formatLocalePath('/locales/$locale/page1.json', 'en-US');
-      expect(results).toEqual('/locales/en-US/page1.json');
+      assume(results).equals('/locales/en-US/page1.json');
     });
     it('substitutes :locale in path template', function () {
       const results = formatLocalePath('/locales/:locale/page1.json', 'en-US');
-      expect(results).toEqual('/locales/en-US/page1.json');
+      assume(results).equals('/locales/en-US/page1.json');
     });
     it('substitutes {locale} in path template', function () {
       const results = formatLocalePath('/locales/{locale}/page1.json', 'en-US');
-      expect(results).toEqual('/locales/en-US/page1.json');
+      assume(results).equals('/locales/en-US/page1.json');
     });
   });
 });
