--- packages/gasket-plugin-intl/jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  collectCoverageFrom: ['lib/**/*.js']
-};
